Show extra totals
TODO the hire cost is not quite identical to the hire cost the optimiser is calculating, even though it's doing pretty much exactly the same thing. I suspect this is something to do with the scale factor.

For now, we can just pull the hire cost out of the fitness components instead.