BugzID: 697
Fixed this case by displaying options for a vessel if its vessel class is in the allowed list. Fixed the exception and some odd behaviour by passing an ElementAssignment object (cloned for editing) to the getAllowedValues() method of the vessel value provider.

Unfortunately another bugfix required removal of the VesselEventVesselsInlineEditor class and factory from the app, thereby re-breaking case 688.