Expanded pricing date options
Create a new object called IPortTimesRecord to replace the existing arrival times arrays and list passed around and to include port visit durations. This is specifically for the pricing API but also feeds into the lower level calculations where this data is created. The IPortTimesRecord can be created with just the cargo slots (when used for pricing) or with the return element (when used in the voyage calculations). Currently there is no real distinction between the two modes, although this is likely to lead to trouble down the line.

IAllocationAnnotations now extend the IPortTimesRecord so it can be used in relevant API calls as it contains the same data.

BugzID: 920