Bug: Optimise spots gets wrong timezone
For spot slots with no port, create two sets of timewindows. First create a UTC month long window.  This is passed into the slot matching code and shifted to be the equivalent local month for each valid target port. Thus midnight 1st Jan in UTC is midnight 1st Jan  local time in Japan or midnight 1st Jan localtime in Canda. This the slot is restricted correctly based on timezone. We then extend the UTC by  -12 to +14 hours to take into account the timezone shifts. This timewindow is used by the time window based constraints and the previous timewindows feed into DES/FOB vessel restrictions.

Finally the export code needs to adjust the final slot window based on the exported port for the slot. This may be set from null or changed from an existing mapping. The timezone needs to shift accordingly.

BugzId: 909