Refactor authentication flow & preference page
This is quite a large commit so here's a brief breakdown of what it
contains:

Application & UpstreamUrlProvider:
* Remove the duplicate login dialog when the datahubStartupCheck is
enabled. It only appears on startup if the hub is online & appears for
everyone
* Remove the authentication dialog pop-up triggered from isUpstreamUrl.
On top of being bad UX design, it makes it hard to test the datahub
preferences because the dialog shows up when the URL is changed. Instead
let users login by going to the datahub preference page

Datahub preference page:
* Disable the login button when the hub is not online. Previously, when
the hub was down, the user could still click the login button and
get a "Can't reach this page" error in the browser
* Clicking the "check connection" button now refreshes the datahub
online status and enables/disables the login button accordingly
* Improved naming of the login button