close shell if oauth is enabled but not working
If oauth is enabled, lingo will open a browser on start up and try to reach `/login/oauth2`.
If oauth is not working because of e.g. a yaml config misconfiguration,
the hub will redirect the browser to `/local/local`. If this happens we
just want to shut down the browser