wip: remove failing hub tests
These tests validate the behaviour of the datahub preference page.
After the change in c2d24180540d4b11ead37b2814e89b9f34d49810 these tests are failing
because of a change in type of the login shell as well as
SWTBot throwing WidgetNotFound issues caused because the tests are run
in a non-UI thread.