Fix: Ballast bonus validation does not update on data change The ballast bonus table in the dialog error did not trigger a validation refresh.
More strictly - it executed the command on the command stack triggering
a full model validation run, but did not use the command handler to
notify the dialog of a change and thus the dialog validation state was
not updated.

This is a big change to switch the all ICellManipulators to use the
command handlers in order to fix this issue.

Resolves #352