Change executors API to create a new thread pool on demand.
Old version used a shared pool and used reflection to clear thread
locals between use. Java 16 blocks access to this API. This new version
also simplifies the code as we no longer need to dispose the thread pool
higher up and we no longer need to maintain maps of thread to objects.